refactor: add shared cluster backend contracts
