+++ npa/src/npa/fleet/tfvars.py
@@ -1,221 +1,7 @@
-"""Render ``terraform.tfvars`` for the k8s-training **root** module + helpers.
+"""Compatibility module alias for the backend-owned mk8s renderer."""
 
-Each fleet cluster is one k8s-training deployment run as a *root* module (the
-recipe references ``../modules`` and embeds its own provider config, so it cannot
-be sourced as a child module). We therefore copy the recipe per cluster, patch
-its provider domain for the target region, and drive it with a rendered
-``terraform.tfvars`` using the recipe's own variable names.
-"""
+import sys
 
-from __future__ import annotations
+from npa.cluster_backends import mk8s_render as _implementation
 
-from pathlib import Path
-from typing import Final
-
-from npa.cluster.gpu_driver import (
-    CANONICAL_RECIPE_CAPABILITIES,
-    GpuDriverStrategyError,
-    RecipeGpuDriverCapabilities,
-    inspect_recipe_declared_variables,
-    inspect_recipe_gpu_driver_capabilities,
-    recipe_driver_tfvars,
-    resolve_gpu_driver_strategy,
-)
-from npa.fleet.mig import (
-    GPU_DEVICE_PLUGIN_VERSION,
-    GPU_DRIVER_VERSION,
-    GPU_GFD_VERSION,
-    GPU_MIG_MANAGER_VERSION,
-    GPU_OPERATOR_VERSION,
-)
-from npa.fleet.spec import ClusterSpec
-
-_EU_DOMAIN = "api.eu.nebius.cloud:443"
-
-MIG_RECIPE_VARIABLES: Final[frozenset[str]] = frozenset(
-    {
-        "gpu_operator_version",
-        "gpu_driver_version",
-        "gpu_device_plugin_version",
-        "gpu_gfd_version",
-        "gpu_mig_manager_version",
-        "gpu_mig_with_reboot",
-        "gpu_operator_rdma_enabled",
-    }
-)
-
-
-def provider_domain(region: str) -> str:
-    """Nebius Terraform provider domain for *region* (EU vs. global)."""
-
-    return _EU_DOMAIN if region.startswith("eu") else "api.nebius.cloud:443"
-
-
-def patch_provider_domain(provider_tf: str, region: str) -> str:
-    """Patch the recipe's hardcoded EU provider domain for *region*."""
-
-    return provider_tf.replace(_EU_DOMAIN, provider_domain(region))
-
-
-def _tfstr(value: str) -> str:
-    escaped = value.replace("\\", "\\\\").replace('"', '\\"')
-    return f'"{escaped}"'
-
-
-def validate_recipe_mig_compatibility(cluster: ClusterSpec, recipe_dir: Path) -> None:
-    """Fail before Terraform when an alternate recipe would drop MIG pins."""
-
-    if not (cluster.mig and cluster.mig.enabled):
-        return
-    declared = inspect_recipe_declared_variables(recipe_dir)
-    missing = sorted(MIG_RECIPE_VARIABLES - declared)
-    if missing:
-        quoted = ", ".join(repr(name) for name in missing)
-        raise GpuDriverStrategyError(
-            f"{recipe_dir} cannot honor RTX PRO 6000 MIG: variables.tf is missing "
-            f"required Terraform variable(s) {quoted}. Use the vendored recipe or "
-            "a compatible --k8s-training-ref/--k8s-training-dir."
-        )
-
-
-def render_tfvars(
-    cluster: ClusterSpec,
-    *,
-    ssh_public_key: str = "",
-    recipe_dir: Path | None = None,
-    recipe_capabilities: RecipeGpuDriverCapabilities | None = None,
-) -> str:
-    """Return ``terraform.tfvars`` for a single cluster (k8s-training root vars).
-
-    tenant/project/region/subnet/iam_token are supplied via ``TF_VAR_*`` env at
-    apply time; everything here maps to the recipe's own variable names.
-    """
-
-    cpu = cluster.cpu_nodes
-    gpu = cluster.gpu_nodes
-    lines: list[str] = [
-        "# Generated by `npa fleet` (k8s-training root module).",
-        f"cluster_name = {_tfstr(cluster.name)}",
-    ]
-    if cluster.resolved_k8s_version():
-        lines.append(f"k8s_version = {_tfstr(cluster.resolved_k8s_version())}")
-    if ssh_public_key:
-        lines.append(f"ssh_public_key = {{ key = {_tfstr(ssh_public_key)} }}")
-
-    lines.append("mk8s_cluster_public_endpoint = true")
-    lines.append("enable_k8s_node_group_sa     = false")
-    lines.append("enable_egress_gateway        = false")
-    lines.append("cpu_nodes_public_ips         = false")
-    lines.append("gpu_nodes_public_ips         = false")
-    lines.append("cpu_nodes_preemptible        = false")
-    lines.append("gpu_nodes_preemptible        = false")
-    mig = cluster.mig
-    mig_enabled = bool(mig and mig.enabled)
-    if recipe_dir is not None:
-        validate_recipe_mig_compatibility(cluster, recipe_dir)
-    lines.append(
-        f"mig_strategy                 = {_tfstr(mig.strategy if mig_enabled else 'none')}"
-    )
-    driver_selection = resolve_gpu_driver_strategy(
-        gpu_nodes=gpu.count if gpu else 0,
-        platform=gpu.platform if gpu else "",
-        preset=gpu.preset if gpu else "",
-        mode=cluster.resolved_gpu_driver_mode(),
-        managed_driver_preset=cluster.managed_driver_preset,
-        enable_gpu_cluster=cluster.resolved_enable_gpu_cluster(),
-        allow_unsafe_nvswitch_operator=cluster.allow_unsafe_nvswitch_operator,
-    )
-    capabilities = recipe_capabilities
-    if capabilities is None and recipe_dir is not None:
-        capabilities = inspect_recipe_gpu_driver_capabilities(recipe_dir)
-    capabilities = capabilities or CANONICAL_RECIPE_CAPABILITIES
-    lines.extend(
-        recipe_driver_tfvars(
-            driver_selection,
-            capabilities,
-            recipe_label=str(recipe_dir or "canonical k8s-training recipe"),
-        )
-    )
-    lines.append(f"custom_driver                = {'true' if mig_enabled else 'false'}")
-    if mig_enabled:
-        lines.append(f"mig_parted_config            = {_tfstr(mig.config)}")
-        lines.append(f"gpu_operator_version         = {_tfstr(GPU_OPERATOR_VERSION)}")
-        lines.append(f"gpu_driver_version           = {_tfstr(GPU_DRIVER_VERSION)}")
-        lines.append(
-            f"gpu_device_plugin_version    = {_tfstr(GPU_DEVICE_PLUGIN_VERSION)}"
-        )
-        lines.append(f"gpu_gfd_version              = {_tfstr(GPU_GFD_VERSION)}")
-        lines.append(
-            f"gpu_mig_manager_version      = {_tfstr(GPU_MIG_MANAGER_VERSION)}"
-        )
-        # RTX PRO Blackwell requires MIG mode and geometry reconciliation across
-        # a driver unload/reboot. Keep this explicit and platform-scoped.
-        lines.append("gpu_mig_with_reboot          = true")
-        # The single-GPU RTX PRO preset has no InfiniBand fabric. Keep the RDMA
-        # driver path off; enabling it adds unnecessary host-module churn.
-        lines.append("gpu_operator_rdma_enabled    = false")
-
-    cpu_count = cpu.count if cpu else 0
-    lines.append(f"cpu_nodes_fixed_count = {cpu_count}")
-    if cpu and cpu.count > 0:
-        lines.append(f"cpu_nodes_platform = {_tfstr(cpu.platform)}")
-        if cpu.preset:
-            lines.append(f"cpu_nodes_preset = {_tfstr(cpu.preset)}")
-        cpu_disk = cpu.disk_size_gib if cpu.disk_size_gib > 0 else 128
-        lines.append(f"cpu_disk_size = {_tfstr(str(cpu_disk))}")
-
-    gpu_count = gpu.count if gpu else 0
-    lines.append(f"gpu_nodes_fixed_count_per_group = {gpu_count}")
-    lines.append(f"gpu_node_groups = {1 if gpu_count > 0 else 0}")
-    if gpu and gpu.count > 0:
-        lines.append(f"gpu_nodes_platform = {_tfstr(gpu.platform)}")
-        if gpu.preset:
-            lines.append(f"gpu_nodes_preset = {_tfstr(gpu.preset)}")
-        if gpu.capacity_block_group:
-            # STRICT is deliberate: AUTO may silently fall back to ordinary
-            # on-demand capacity when the named capacity block is exhausted.
-            lines.append(
-                'gpu_nodes_reservation_policy = { policy = "STRICT", '
-                f"reservation_ids = [{_tfstr(gpu.capacity_block_group)}] }}"
-            )
-        lines.append(
-            f"gpu_disk_size = {_tfstr(str(cluster.resolved_gpu_disk_size_gib()))}"
-        )
-
-    lines.append(
-        f"enable_gpu_cluster = {'true' if cluster.resolved_enable_gpu_cluster() else 'false'}"
-    )
-    if cluster.infiniband_fabric:
-        lines.append(f"infiniband_fabric = {_tfstr(cluster.infiniband_fabric)}")
-
-    lines.append(
-        f"enable_filestore = {'true' if cluster.enable_filestore else 'false'}"
-    )
-    # Always emit existing_filestore: the recipe's variable defaults to null, and
-    # its filesystem.tf branches on `== ""` / `!= ""`. Left null, enable_filestore
-    # would try to READ a nonexistent filesystem (data source with no id) instead
-    # of creating one. Empty string routes to the create branch.
-    lines.append(f"existing_filestore = {_tfstr(cluster.existing_filestore)}")
-    lines.append(
-        f"filestore_disk_size_gibibytes = {cluster.filestore_disk_size_gibibytes}"
-    )
-    lines.append(f"filestore_mount_path = {_tfstr(cluster.filestore_mount_path)}")
-    lines.append(f"filestore_mount_tag = {_tfstr(cluster.filestore_mount_tag)}")
-    lines.append(
-        'filesystem_csi = { chart_version = "0.1.6", namespace = "kube-system", '
-        "make_default_storage_class = true, previous_default_storage_class_name = "
-        '"compute-csi-default-sc" }'
-    )
-
-    # Keep the fleet cheap and quiet: no observability/logging/ray/gatekeeper.
-    lines.append("enable_nebius_o11y_agent = false")
-    lines.append("enable_grafana           = false")
-    lines.append("enable_prometheus        = false")
-    lines.append("collectK8sClusterMetrics = false")
-    lines.append("enable_kuberay_cluster   = false")
-    lines.append("enable_kuberay_service   = false")
-    lines.append("enable_opa_gatekeeper    = false")
-    # loki has no default in the recipe; it must be set explicitly.
-    lines.append("loki = { enabled = false, replication_factor = 1 }")
-
-    return "\n".join(lines) + "\n"
+sys.modules[__name__] = _implementation
